- feat: 장바구니 API/서비스 구축 및 CART 기반 주문 생성 연동

+++ src/main/java/com/jkomerce/store/controller/ItemController.java
@@ -26,7 +26,7 @@ public class ItemController {
 
     @GetMapping("/detail/{id}")
     @ResponseBody
-    public ItemDTO getDetailPage(@PathVariable int id) {
+    public ItemDTO getDetailPage(@PathVariable Long id) {
         return itemService.getItemById(id);
     }
